Delete outdated pipelines/utils.py
Reorganized and fixed tests.
